requirements.txt
fastapi
uvicorn
prometheus-client
pyyaml

requests
